requirements.txt
discord.py
python.decouple
